requirements.txt
Django~=4.2.5
gunicorn~=21.2.0
django-heroku
whitenoise
dj-database-url
cryptography~=41.0.4
asgiref~=3.7.2
sqlparse~=0.4.4
psycopg2~=2.9.8
requests~=2.31.0
django-allauth~=0.57.0
PyJWT~=2.8.0
certifi~=2023.7.22
oauthlib~=3.2.2
idna~=3.4
urllib3~=2.0.5
pycparser~=2.21
django-bootstrap-v5